backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-multipart>=0.0.12,<1.0.0
Pillow>=12.0.0,<13.0.0
pytesseract>=0.3.13,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.5.2,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
openai>=1.54.0,<2.0.0
slowapi>=0.1.9,<1.0.0
tenacity>=8.2.0,<10.0.0
